Moved experimenter_factory to own file
