+++ backend/users/experimenter.py
@@ -746,53 +746,3 @@ class Experimenter(User):
             type="SET_FILTERS", description="Successfully changed filters."
         )
         return MessageDict(type="SUCCESS", data=success)
-
-
-async def experimenter_factory(
-    offer: RTCSessionDescription, id: str, hub: _h.Hub
-) -> tuple[RTCSessionDescription, Experimenter]:
-    """Instantiate connection with a new Experimenter based on WebRTC `offer`.
-
-    Instantiate new modules.experimenter.Experimenter, handle offer using
-    modules.connection.connection_factory and set connection for the Experimenter.
-
-    This sequence must be donne for all experimenters.  Instantiating an Experimenter
-    directly will likely lead to problems, since it wont have a Connection.
-
-    Parameters
-    ----------
-    offer : aiortc.RTCSessionDescription
-        WebRTC offer for building the connection to the client.
-    id : str
-        Unique identifier for Experimenter.
-    hub : modules.hub.Hub
-        Hub the Experimenter will be part of.  Used for api calls.
-
-    Returns
-    -------
-    tuple with aiortc.RTCSessionDescription, modules.experimenter.Experimenter
-        WebRTC answer that should be send back to the client and Experimenter
-        representing the client.
-    """
-    experimenter = Experimenter(id, hub)
-    filter_api = FilterAPI(experimenter)
-    log_name_suffix = f"E-{id}"
-
-    if hub.config.experimenter_multiprocessing:
-        answer, connection = await connection_subprocess_factory(
-            offer,
-            experimenter.handle_message,
-            log_name_suffix,
-            hub.config,
-            [],
-            [],
-            filter_api,
-            (False, '')
-        )
-    else:
-        answer, connection = await connection_factory(
-            offer, experimenter.handle_message, log_name_suffix, [], [], filter_api, (False, '')
-        )
-
-    experimenter.set_connection(connection)
-    return (answer, experimenter)
